#169: `/bounty post` flow ends in replying to ephemeral message - fix crash on delete channel in guild without company row
